+++ hutool-core/src/main/java/com/xiaoleilu/hutool/date/DatePattern.java
@@ -43,4 +43,5 @@ public class DatePattern {
 	public final static String JDK_DATETIME_PATTERN = "EEE MMM dd HH:mm:ss zzz yyyy";
 	/** HTTP头中日期时间格式 {@link FastDateFormat}：EEE MMM dd HH:mm:ss zzz yyyy */
 	public final  static FastDateFormat JDK_DATETIME_FORMAT = FastDateFormat.getInstance(JDK_DATETIME_PATTERN);
+	
 }