+++ hutool-core/src/main/java/com/xiaoleilu/hutool/lang/PatternPool.java
@@ -6,22 +6,25 @@ import com.xiaoleilu.hutool.util.ReUtil;
 
 /**
  * 常用正则表达式集合
+ * 
  * @author Looly
  *
  */
 public class PatternPool {
+	
 	/** 英文字母 、数字和下划线 */
 	public final static Pattern GENERAL = Pattern.compile("^\\w+$");
 	/** 数字 */
 	public final static Pattern NUMBERS = Pattern.compile("\\d+");
-	/**单个中文汉字 */
+	/** 单个中文汉字 */
 	public final static Pattern CHINESE = Pattern.compile(ReUtil.RE_CHINESE);
-	/**中文汉字 */
+	/** 中文汉字 */
 	public final static Pattern CHINESES = Pattern.compile(ReUtil.RE_CHINESES);
 	/** 分组 */
 	public final static Pattern GROUP_VAR = Pattern.compile("\\$(\\d+)");
 	/** IP v4 */
-	public final static Pattern IPV4 = Pattern.compile("\\b((?!\\d\\d\\d)\\d+|1\\d\\d|2[0-4]\\d|25[0-5])\\.((?!\\d\\d\\d)\\d+|1\\d\\d|2[0-4]\\d|25[0-5])\\.((?!\\d\\d\\d)\\d+|1\\d\\d|2[0-4]\\d|25[0-5])\\.((?!\\d\\d\\d)\\d+|1\\d\\d|2[0-4]\\d|25[0-5])\\b");
+	public final static Pattern IPV4 = Pattern.compile(
+			"\\b((?!\\d\\d\\d)\\d+|1\\d\\d|2[0-4]\\d|25[0-5])\\.((?!\\d\\d\\d)\\d+|1\\d\\d|2[0-4]\\d|25[0-5])\\.((?!\\d\\d\\d)\\d+|1\\d\\d|2[0-4]\\d|25[0-5])\\.((?!\\d\\d\\d)\\d+|1\\d\\d|2[0-4]\\d|25[0-5])\\b");
 	/** 货币 */
 	public final static Pattern MONEY = Pattern.compile("^(\\d+(?:\\.\\d+)?)$");
 	/** 邮件 */
@@ -46,4 +49,107 @@ public class PatternPool {
 	public final static Pattern PLATE_NUMBER = Pattern.compile("^[京津沪渝冀豫云辽黑湘皖鲁新苏浙赣鄂桂甘晋蒙陕吉闽贵粤青藏川宁琼使领A-Z]{1}[A-Z]{1}[A-Z0-9]{4}[A-Z0-9挂学警港澳]{1}$");
 	/** MAC地址正则 */
 	public static final Pattern MAC_ADDRESS = Pattern.compile("((?:[A-F0-9]{1,2}[:-]){5}[A-F0-9]{1,2})|(?:0x)(\\d{12})(?:.+ETHER)", Pattern.CASE_INSENSITIVE);
+
+	// -------------------------------------------------------------------------------------------------------------------------------------------------------------------
+	/** Pattern池 */
+	private static final SimpleCache<RegexWithFlag, Pattern> POOL = new SimpleCache<>();
+	
+	/**
+	 * 先从Pattern池中查找正则对应的{@link Pattern}，找不到则编译正则表达式并入池。
+	 * @param regex 正则表达式
+	 * @return {@link Pattern}
+	 */
+	public static Pattern get(String regex) {
+		return get(regex, 0);
+	}
+
+	/**
+	 * 先从Pattern池中查找正则对应的{@link Pattern}，找不到则编译正则表达式并入池。
+	 * @param regex 正则表达式
+	 * @param flags 正则标识位集合 {@link Pattern}
+	 * @return {@link Pattern}
+	 */
+	public static Pattern get(String regex, int flags) {
+		final RegexWithFlag regexWithFlag = new RegexWithFlag(regex, flags);
+		
+		Pattern pattern = POOL.get(regexWithFlag);
+		if(null == pattern){
+			pattern = Pattern.compile(regex, flags);
+			POOL.put(regexWithFlag, pattern);
+		}
+		return pattern;
+	}
+	
+	/**
+	 * 移除缓存
+	 * @param regex 正则
+	 * @param flags 标识
+	 * @return 移除的{@link Pattern}，可能为{@code null}
+	 */
+	public static Pattern remove(String regex, int flags){
+		return POOL.remove(new RegexWithFlag(regex, flags));
+	}
+	
+	/**
+	 * 清空缓存池
+	 */
+	public static void clear(){
+		POOL.clear();
+	}
+
+	//---------------------------------------------------------------------------------------------------------------------------------
+	/**
+	 * 正则表达式和正则标识位的包装
+	 * @author Looly
+	 *
+	 */
+	private static class RegexWithFlag {
+		private String regex;
+		private int flag;
+
+		/**
+		 * 构造
+		 * @param regex 正则
+		 * @param flag 标识
+		 */
+		public RegexWithFlag(String regex, int flag) {
+			this.regex = regex;
+			this.flag = flag;
+		}
+
+		@Override
+		public int hashCode() {
+			final int prime = 31;
+			int result = 1;
+			result = prime * result + flag;
+			result = prime * result + ((regex == null) ? 0 : regex.hashCode());
+			return result;
+		}
+
+		@Override
+		public boolean equals(Object obj) {
+			if (this == obj) {
+				return true;
+			}
+			if (obj == null) {
+				return false;
+			}
+			if (getClass() != obj.getClass()) {
+				return false;
+			}
+			RegexWithFlag other = (RegexWithFlag) obj;
+			if (flag != other.flag) {
+				return false;
+			}
+			if (regex == null) {
+				if (other.regex != null) {
+					return false;
+				}
+			} else if (!regex.equals(other.regex)) {
+				return false;
+			}
+			return true;
+		}
+
+	}
 }